pyproject.toml
[tool.poetry]
name = "setup_creator"
version = "0.1.0"
description = ""
authors = ["K0GiHF9S <kidoairakusyunkasyuutou@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
setup_creator = "setup_creator.__main__:main"

[tool.poetry.dependencies]
python = "^3.9"
toml = "^0.10.2"
msl-loadlib = "^0.8.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
pytest-cov = "^2.11.1"
pytest-datadir = "^1.3.1"
autopep8 = "^1.5.7"
coverage = {extras = ["toml"], version = "^5.5"}
pytest-html = "^3.1.1"

[tool.coverage]
    [tool.coverage.run]
    branch = true
    source = ["."]
    omit = [".venv/*", "tests/*"]

    [tool.coverage.report]
    exclude_lines = ["pragma: no cover","if __name__ == .__main__.:"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
